+++ ta4j-core/src/main/java/org/ta4j/core/indicators/helpers/CombineIndicator.java
@@ -23,12 +23,12 @@
  */
 package org.ta4j.core.indicators.helpers;
 
+import java.util.function.BinaryOperator;
+import java.util.function.Function;
+
 import org.ta4j.core.Indicator;
 import org.ta4j.core.indicators.CachedIndicator;
 import org.ta4j.core.num.Num;
-
-import java.util.function.BinaryOperator;
-import java.util.function.Function;
 
 /**
  * Combine indicator.
