+++ ta4j-core/src/test/java/org/ta4j/core/indicators/adx/PlusDMIndicatorTest.java
@@ -33,7 +33,6 @@ import org.junit.Test;
 import org.ta4j.core.Bar;
 import org.ta4j.core.Indicator;
 import org.ta4j.core.indicators.AbstractIndicatorTest;
-import org.ta4j.core.indicators.adx.PlusDMIndicator;
 import org.ta4j.core.mocks.MockBar;
 import org.ta4j.core.mocks.MockBarSeries;
 import org.ta4j.core.num.Num;
